Comment duplication bug fixed

+++ android/src/main/java/com/sc/mtaa_safi/feed/comments/CommentAdapter.java
@@ -26,7 +26,7 @@ public class CommentAdapter extends RecyclerViewCursorAdapter<CommentAdapter.Vie
     public void onBindViewHolder(ViewHolder holder, Cursor c) {
         holder.mTextView.setText(c.getString(c.getColumnIndex(Contract.Comments.COLUMN_CONTENT)));
         holder.mUserNameView.setText(c.getString(c.getColumnIndex(Contract.Comments.COLUMN_USERNAME)));
-        holder.mTimeView.setText(PrefUtils.getElapsedTime(c.getInt(c.getColumnIndex(Contract.Comments.COLUMN_TIMESTAMP))));
+        holder.mTimeView.setText(PrefUtils.getElapsedTime(c.getLong(c.getColumnIndex(Contract.Comments.COLUMN_TIMESTAMP))));
     }
 
     public static class ViewHolder extends RecyclerView.ViewHolder {
